Move controllers to citizen hub directory
